integra mudanças na view
