+++ src/pages/Inscricao/Interfaces.tsx
@@ -1,7 +1,12 @@
+export interface IFile {
+  id: number;
+  fileData: File;
+}
+
 export interface IInscricaoData {
-  historicosGraduacao: FileList | null;
-  historicosPosGraduacao: FileList | null;
-  producoesCientificas: FileList | null;
+  historicosGraduacao: IFile[];
+  historicosPosGraduacao: IFile[];
+  producoesCientificas: IFile[];
   enade: string;
   checkboxes: boolean[];
 }
